requirements.txt
gradio==4.44.1
torch
torchvision
opencv-python-headless
numpy
hydra-core
iopath
Pillow